pyproject.toml
[tool.poetry]
name = "users-mcr"
version = "0.1.0"
description = ""
authors = ["Anastasia Druzianova <90915361+markbeth@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = {extras = ["standard"], version = "^0.27.1"}
greenlet = "^3.0.3"
gunicorn = "^21.2.0"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.2"
bcrypt = "^4.1.2"
fastapi-cache2 = "^0.2.1"
prometheus-client = "^0.20.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
python-json-logger = "^2.0.7"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
fastapi-mail = "^1.4.1"
jinja2 = "^3.1.3"
fastapi-versioning = "^0.10.0"
pydantic-settings = "^2.2.1"
sentry-sdk = "^1.41.0"
sqladmin = "^0.16.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "app"
asyncio_mode = "auto"